Mean vertex calibration (#4638)
* First implementation of MeanVertex calibration

In the direction of the moving average.

Vectors for SMA

Compiling, but I did not check if it works

typos

Processing MeanVertex.

* First bunch of comments by Ruben (a couple left, for next week :-) )

Following review to PR.

following review

clang-format

* Using ConfigurableParams

clang

* remove mTmpMVobjDqTimeEnd, and fixes

* Fix determination of startValidity

clang

* Using Bracket

Using Bracket and changes related to modifications of Utils in dev

clang

+++ Detectors/TOF/calibration/src/LHCClockCalibrator.cxx
@@ -81,7 +81,7 @@ void LHCClockCalibrator::finalizeSlot(Slot& slot)
             << c->getEntries() << " entries";
   std::vector<float> fitValues;
   float* array = &c->histo[0];
-  int fitres = fitGaus(c->nbins, array, -(c->range), c->range, fitValues);
+  double fitres = fitGaus(c->nbins, array, -(c->range), c->range, fitValues);
   if (fitres >= 0) {
     LOG(INFO) << "Fit result " << fitres << " Mean = " << fitValues[1] << " Sigma = " << fitValues[2];
   } else {
